PDB_1115
- 카테고리 joblist와 연결
- 연결시켜 등록은 아직 못함
- jobList jno를 int로 수정
- login css 변경
- 아이디 찾기 경로 설정

+++ src/main/java/com/ds/antddun/dto/JobListDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class JobListDTO {
-    private Long jno;
+    private int jno;
     private String job;
 //    private Member member;
