+++ src/main/java/com/ds/antddun/entity/JobList.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 public class JobList {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "JOBLIST_SEQ_GEN")
-    private Long jno;
+    private int jno;
 
     private String job;
