+++ src/main/java/com/ds/antddun/repository/JobListRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface JobListRepository extends JpaRepository<JobList, Long> {
+public interface JobListRepository extends JpaRepository<JobList, Integer> {
 
 
 }
